refactor: Move Ui and Parameter out of model package
